Restructure code, add makefile and vscode tools
